Override existing class rather than create a new one
