+++ scorecard/profile_data/indicators/series.py
@@ -12,18 +12,3 @@ class SeriesIndicator():
             "last_year": years[0] if len(years) > 0 else None,
             "formula": cls.formula,
         }
-
-class SeriesIndicatorUIFW():
-
-    @classmethod
-    def get_muni_specifics(cls, api_data):
-        results = api_data.results
-        years = api_data.uifw_years
-
-        return {
-            "result_type": cls.result_type,
-            "values": cls.get_values(years, results),
-            "ref": api_data.references[cls.reference],
-            "last_year": years[0] if len(years) > 0 else None,
-            "formula": cls.formula,
-        }
